requirements.txt
asgiref==3.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.7
django-allauth==0.50.0
django-ckeditor==6.4.0
django-cors-headers==3.12.0
django-filter==21.1
django-js-asset==2.0.0
django-taggit==3.0.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.7
oauthlib==3.2.0
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
urllib3==1.26.9
zipp==3.8.0
